Refactored functions into hepers and moved fils around
